Refactor editor to its own component
