+++ frontend/pages/auth/portal/dashboard/post/[year]/[month]/[slug]/index.js
@@ -1,0 +1,11 @@
+import Layout from '../../../../../../../../components/DashboardLayout/dashboardLayout';
+
+const Index = ({ auth, dispatch }) => {
+    return (
+        <Layout selectedCategory={"None"}>
+            <h1>Post</h1>
+        </Layout>
+    );
+};
+
+export default Index;
